requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
Werkzeug==3.0.1
markdown==3.5.1
bleach==6.1.0
